Move envvar test to sidecar tests
